feat: add better log styling and calculate reading time
Previously, I was going to use @tailwindcss/typography for styling, but it caused issues with .mdx components. Since I plan to make heavy use of .mdx for illustrations, I want to keep that open. It's not like typography does that much for me anyway. I got it mostly figured out.